setup typescript
