+++ .parcelrc
@@ -1,0 +1,13 @@
+{
+  "extends": [
+    "@parcel/config-default"
+  ],
+  "reporters": [
+    "...",
+    "parcel-reporter-static-files-copy"
+  ],
+  "resolvers": [
+    "parcel-resolver-ignore",
+    "..."
+  ]
+}
