+++ package.json
@@ -2,20 +2,26 @@
   "name": "rtp-start-app",
   "version": "0.1.0",
   "private": true,
+  "source": "src/index.html",
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.7.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "BUILD_PATH='./dist' react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "generate": "",
+    "start": "npm run generate && parcel --no-cache",
+    "build": "npm run generate && tsc & parcel build",
+    "lint": "eslint . -c .eslintrc.js --ext .ts,.tsx",
+    "lint:fix": "eslint . -c .eslintrc.js --ext .ts,.tsx --fix",
+    "test": "vitest",
+    "coverage": "vitest --coverage"
   },
   "eslintConfig": {
     "extends": [
@@ -34,5 +40,18 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.0",
+    "dotenv": "^16.4.5",
+    "parcel": "^2.12.0",
+    "parcel-reporter-static-files-copy": "^1.5.3",
+    "parcel-resolver-ignore": "^2.2.0",
+    "process": "^0.11.10",
+    "rimraf": "^6.0.1",
+    "typescript": "^5.6.2",
+    "vite": "^5.4.8",
+    "vite-tsconfig-paths": "^5.0.1"
   }
 }
