+++ static/global.css
@@ -1,0 +1,5 @@
+
+body {
+    margin: 0px;
+    padding: 0px;
+}
